step 1: move new logger into separate dir
