+++ http/server/customlogging/custom_logging_suite_test.go
@@ -1,0 +1,13 @@
+package customlogging
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestCustomLogging(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Custom Logging Suite")
+}
