+++ http/server/recovery/recovery.go
@@ -11,7 +11,7 @@ import (
 type PanicHandlerFunc func(logger slog.Logger, w http.ResponseWriter, r *http.Request, err interface{})
 
 func defaultPanicHandler(logger slog.Logger, w http.ResponseWriter, r *http.Request, err interface{}) {
-	attributes := logging.BuildAttributes(r.Context(), "", r, map[string]interface{}{"err": err})
+	attributes := logging.BuildAttributes(r.Context(), r, "error", err)
 	logger.ErrorContext(r.Context(), "Panic occurred", attributes...)
 	http.Error(w, "Internal Server Error", http.StatusInternalServerError)
 }
